irqchip/sifive-plic: Unset bit in x280 global interrupts register in handler
blackhole l2cpu complex has 4 x280 global interrupts registers
that are level triggered. The bits in these need to be unset
while handling the interrupt.

The device side (i.e the person trigerring the interrupt) could do it
but it seems to be too slow/late and results in some spurious
interrupts. So we do it here instead

Signed-off-by: Anirudh Srinivasan <asrinivasan@tenstorrent.com>